Changed semantics of PortControl::send().
Changed PortControl::send() so that it no longer checks
to see if there is space for the packet being sent.  It is
now the responsibility of the sending class to explicitly
call PortControl::spaceToSend() to check if there is space
available.  This allows for high level and/or best case type
models to ignore the output buffer size if desired.