Change to seeding for random number generator in TrafficGen.
When setting a seed in TrafficGen, the random number generator was recreated, but the distribution was not, leading the changes to be ignored.  Seeding will now create both a new RNG and a new distribution.  A change was made to allow reseeding the underlying RNG, which should be tested and used if it works.