Added new Dragonfly models, ability to specify port parameters by logical grouping and new shift_test test.
New Dragonfly models are currently called Dragonfly2, but will become the only dragonfly model within the next few months.  They are more flexible and allow for unconnected ports on routers.  Logical port groups allows a topology to group the ports in a router and have them configured differently that other groups of ports.