Added adaptive-local routing scheme to dragonfly2 topology.
The adaptive-local scheme chooses between upto two direct and two
indirect routes when injected into the network.  The decision is
made using only infomration local to the router.  There is also a
decision made between up to two direct routes when an adaptively
routed packet enters the midgroup and needs to be routed to the
final group.