Added relative global route mode.
The new relative global route mode interprets the global
link map differently.  Numbers in the link map are taken
as relative offsets from the current group, instead of as
an absolute number with the current group removed.  The
translation from global link map entry to raw group number
and global slice is the same for both modes.  However, the
raw group numbers are converted to actual group number
differently.