Removed the VN autoconfigure functionality from hr_router.
The routers default to 2 VNs if the number is not specified.  This
should work for most simple simulations.  New features have made it
too complex to continue to support autoconfiguring based on the
requested VNs from the endpoints.  The endpoints will still need to
request the number of VNs they want so that VN remapping will
continue to work.