Some fixes and enhancements for python modules in merlin and ember.
Added the ability for the merlin System object to allocate in blocks greater than
one.  This was needed to enable multinic configurations.

Added more flexibility to enabling statistics through the template objects.

Also fixed a VN bug in linkcontrol.  Linkcontrol now sets the VN field in
the request object to be the VN used in send().