[vanadis] Implement the missing Zicntr instructions.
Chapter 8 in
https://github.com/riscv/riscv-isa-manual/releases/download/riscv-isa-release-f797123-2024-06-27/riscv-unprivileged.pdf
details the RISC-V counter interface.

There are 3 bespoke counters detailed in the Zicntr extensions in 8.1, and up to
29 more user-programmable counters detailed in the Zihpm extenstion in 8.2. This
commit provides space for all 32 counters in the register file, but only
currently implements the three Zicntr counters.

The implementation consists of a few changes.

1. I have extended the register file structure with 32, 64 bit counters, and
   `increment` and `get` members in order to update and read those counters.
2. I have added a new instruction, `Zicntr::VanadisReadCounterInstruction` in
   order to read those counters. This instruction can be tailored for any of the
   3 `Zicntr`s (implemented as tag types in `inst/zicntr.h`), XLEN=64 or 32, and
   with or without the `[H]` extension.
3. I have added decoding to `decoder/vriscv64decoder.h` to handle `RDCYCLE`,
   `RDTIME`, and `RDINSTRET`. Because the `[H]` and XLEN=32 versions are only
   available in `riscv32`, the `riscv64` decoder simply injects a decoding
   failure if those occur.
4. I have extended the core's `tick` to update the three `Zicntr`s.
5. I have added a small misc test example to demonstrate usage.

I have not implemented any functionality for programming the 29 `Zihpm`
counters, nor have I added any special decoding to read these registers.

I have decided that the `Zicntr` instructions will be processed by the
arithmetic functional unit. I don't know if this is appropriate, but it should
manage any register port contention properly and seemed like the most expedient
solution.

I have removed the `cycle_count` and the `getCycleCount` from the decoder base
class, as it is no longer being used. The current cycle is still passed to the
decoder `tick()` even though it is _also_ still unused.