merlin/bridge: load user subcomponent for networkIF
pymerlin expects to override a SimpleNetwork subcomponent with its own
network subcomponent (e.g., LinkControl). When using a Bridge, this
subcomponent is ignored and link initialization crashes on
SST::Link::sendUntimedData when trying to access a null rtr object. This
patch extends the Bridge component to use the user subcomponent if the
subcomponent slot is populated and if not then fall back to using an
anonymous subcomponent as before.