miranda: check request operation instead of using dynamic_cast
For my simulation, approximately 15-20% of CPU time is spent resolving one
dynamic_cast in miranda. This dynamic cast does not appear to be necessary
because mirandaCPU dispatches GeneratorRequests based on the request's
operation.

This patch replaces the dynamic_cast with an operation check and adds an
assertion on the MemoryOpRequest's operation type. The assertion ensures
a MemoryOpRequest is either READ or WRITE to avoid potential incorrect
static_cast'ing into CustomOpRequests.